Separating InstancesList and InstancesListPage
