+++ src/types/Instance.ts
@@ -1,0 +1,8 @@
+import { IRow } from "@patternfly/react-table";
+
+export interface Instance extends IRow {
+  id: string;
+  name: string;
+  status: string;
+  submitted_at: string;
+}
